+++ srcs/exec/mini_humans.c
@@ -6,18 +6,36 @@
 /*   By: dpaes-so <dpaes-so@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/05 12:27:43 by dpaes-so          #+#    #+#             */
-/*   Updated: 2025/05/13 15:28:48 by dpaes-so         ###   ########.fr       */
+/*   Updated: 2025/05/13 17:38:57 by dpaes-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incs/mini_header.h"
 
+void	here_loop(int j, t_cmd *cmds,int fd[2])
+{
+	int		i;
+	char	*str;
+
+	while (1)
+	{
+		i = 0;
+		str = readline("> ");
+		if (!str || !ft_strcmp(str, cmds->redir[j].value))
+		{
+			free(str);
+			break ;
+		}
+		while (str[i])
+			write(fd[1], &str[i++], 1);
+		write(fd[1], "\n", 1);
+		free(str);
+	}
+}
 int	here_doc(t_pipe pipex, t_cmd *cmds, int j, t_mini *mini)
 {
-	char	*str;
-	int		fd[2];
-	int		i;
-	int		pid;
+	int	fd[2];
+	int	pid;
 
 	(void)pipex;
 	pipe(fd);
@@ -27,62 +45,17 @@ int	here_doc(t_pipe pipex, t_cmd *cmds, int j, t_mini *mini)
 	if (pid == 0)
 	{
 		choose_signal(3);
-		ft_printf("im after\n");
-		while (1)
-		{
-			i = 0;
-			str = readline("> ");
-			if (!str || !ft_strcmp(str, cmds->redir[j].value))
-			{
-				free(str);
-				break ;
-			}
-			while (str[i])
-				write(fd[1], &str[i++], 1);
-			write(fd[1], "\n", 1);
-			free(str);
-		}
-		ft_printf("BRO MATEIME\n");
-		exit_childprocess(mini,0);
+		here_loop(j, cmds,fd);
+		exit_childprocess(mini, 0);
 	}
 	else
 		wait(NULL);
-	ft_printf("olha retornei isto\n");
 	close(fd[1]);
 	return (fd[0]);
 }
 
-// int	here_doc(t_pipe pipex, t_cmd *cmds)
-// {
-// 	char	*str;
-// 	int		fd[2];
-// 	int		i;
-
-// 	(void)pipex;
-// 	pipe(fd);
-// 	choose_signal(3);
-// 	while (1)
-// 	{
-// 		i = 0;
-// 		str = readline("> ");
-// 		if (!str || !ft_strncmp(str, cmds->redir[0].value,
-// 				ft_strlen(cmds->redir[0].value)))
-// 		{
-// 			free(str);
-// 			break ;
-// 		}
-// 		while (str[i])
-// 			write(fd[1], &str[i++], 1);
-// 		write(fd[1], "\n", 1);
-// 		free(str);
-// 	}
-// 	close(fd[1]);
-// 	return (fd[0]);
-// }
-
 void	first_child(t_mini *mini, t_cmd cmds)
 {
-	ft_printf("IM first\n");
 	do_redirect(&cmds, mini);
 	if (!cmds.cmd)
 		exit_childprocess(mini, 0);
@@ -166,7 +139,6 @@ void	solo_child(t_mini *mini, t_cmd cmds)
 	pid = fork();
 	if (pid == 0)
 	{
-		ft_printf("IM HEdddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddRE\n");
 		choose_signal(2);
 		do_redirect(&cmds, mini);
 		if (!cmds.cmd)
